esp8266: increased xmodem timeout to handle web get lags

+++ msx/apps/xrecv2/serial.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <system_vars.h>
 
-bool wait_for_byte(uint8_t period) __z88dk_fastcall {
+bool wait_for_byte(uint16_t period) __z88dk_fastcall {
   const int16_t timeout = ((int16_t)JIFFY) + period;
 
   while ((timeout - ((int16_t)JIFFY) >= 0) && !fossil_rs_in_stat())
