+++ msx/apps/xrecv2/serial.h
@@ -5,6 +5,6 @@
 
 #define VDP_FREQUENCY 50
 
-extern bool wait_for_byte(uint8_t period) __z88dk_fastcall;
+extern bool wait_for_byte(uint16_t period) __z88dk_fastcall;
 
 #endif
